+++ probe.go
@@ -77,6 +77,7 @@ func handleProbe(scrapers []collector.Scraper, logger *slog.Logger) http.Handler
 			collector.SetLockWaitTimeout(*exporterLockTimeout),
 			collector.SetSlowLogFilter(*slowLogFilter),
 			collector.SetQueryTimeout(time.Duration(*exporterQueryTimeout)*time.Second),
+			collector.SetMaxOpenConns(*exporterMaxOpenConns),
 		))
 
 		h := promhttp.HandlerFor(registry, promhttp.HandlerOpts{})